+++ app/containers/App.js
@@ -6,15 +6,13 @@ import { Provider } from 'react-redux';
 // import Login from './Login';
 import UpcomingShifts from './UpcomingShifts';
 import SettingsTabs from './SettingsTabs';
-import TeamSettings from './TeamSettings';
-import EditTeamSettings from './EditTeamSettings';
-import EditUserSettings from './EditUserSettings';
-import Footer from './Footer';
+import EditTeamSettings from '../settings/EditTeamSettings';
+import EditUserSettings from '../settings/EditUserSettings';
 import Dashboard from './Dashboard';
+import Notifications from './Notifications';
 import { configureStore } from '../redux/utils/store';
 
 import { NativeRouter, Route, Link } from 'react-router-native';
-// import Home from './UpcomingShifts';
 
 type Props = any;
 
@@ -26,8 +24,9 @@ export default class App extends Component<Props> {
       <Provider store={store}>
         <NativeRouter>
           <Route exact path="/" component={Dashboard} />
-          <Route exact path={'/settings'} component={SettingsTabs} />
+          <Route exact path={'/notify'} component={Notifications} />
           <Route exact path={'/shifts'} component={UpcomingShifts} />
+          <Route exact path={'/settings'} component={SettingsTabs} />
           <Route exact path={'/editteamsettings'} component={EditTeamSettings} />
           <Route exact path={'/editusersettings'} component={EditUserSettings} />
         </NativeRouter>
