+++ app/containers/AppFooter.js
@@ -1,30 +1,42 @@
 import React, { Component } from 'react';
 import { Container, Header, Content, Footer, FooterTab, Button, Icon, Text } from 'native-base';
 import { NativeRouter, Route, Link } from 'react-router-native';
-export default class FooterTabsIconTextExample extends Component {
+import { View } from 'react-native';
+
+export default class AppFooter extends Component {
   render() {
     return (
         <Footer>
           <FooterTab>
-            <Button vertical active>
+            <Button>
+              <Link to={'/'}>
               <Icon name="apps" />
+              </Link>
               <Text uppercase={false}>Dashboard</Text>
             </Button>
-            <Button vertical>
+          </FooterTab>
+          <FooterTab>
+            <Button>
+              <Link to={'/notify'}>
               <Icon active name="chatboxes" />
+              </Link>
               <Text uppercase={false}>Notifications</Text>
             </Button>
-            <Button vertical>
+          </FooterTab>
+          <FooterTab>
+            <Button>
               <Link to={'/shifts'}>
               <Icon name="calendar" />
               </Link>
               <Text uppercase={false}>Calendar</Text>
             </Button>
-            <Button vertical>
-              <Link to={'/settings'}>
-              <Icon name="cog" />
-              </Link>
-              <Text uppercase={false}>Settings</Text>
+          </FooterTab>
+          <FooterTab>
+            <Button>
+            <Link to={'/settings'}>
+            <Icon name="cog" />
+            </Link>
+            <Text uppercase={false}>Settings</Text>
             </Button>
           </FooterTab>
         </Footer>
