+++ app/settings/TeamSettings.js
@@ -1,6 +1,7 @@
 // @flow
 
 import React, { Component } from 'react';
+import { View } from 'react-native';
 import { Agenda } from 'react-native-calendars';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
@@ -9,7 +10,20 @@ import moment from 'moment';
 import { getAllShifts } from '../redux/actions/shifts';
 import { renderItem, renderEmptyDate, rowHasChanged } from '../components/AgendaItems';
 
-import { Container, Content, Header, Left, Right, Body, Title, Text, Button, Card, CardItem } from 'native-base';
+import {
+  Container,
+  Icon,
+  Content,
+  Header,
+  Left,
+  Right,
+  Body,
+  Title,
+  Text,
+  Button,
+  Card,
+  CardItem,
+} from 'native-base';
 
 import { Link } from 'react-router-native';
 
@@ -17,40 +31,33 @@ type Props = {
   getAllShifts: () => void,
 };
 
-class UserSettings extends Component<Props> {
+class TeamSettings extends Component<Props> {
   componentWillMount() {
     this.props.getAllShifts();
   }
 
   render() {
-
     return (
       <Container>
-      <Content contentContainerStyle ={{margin: 25}}>
-        <Card>
-        <CardItem>
-        <Text>
-        Name: Rikki Kendall
-        </Text>
-        </CardItem>
-        <CardItem>
-        <Text>
-        Email: yrk3@duke.edu
-        </Text>
-        </CardItem>
-        <CardItem>
-        <Text>
-        Phone: 5593266408
-        </Text>
-        </CardItem>
-        </Card>
-        <Button block style= {{marginTop: 40, margin: 50}}>
-        <Link to={'/editusersettings'}>
-        <Text> Edit User Information </Text>
-        </Link>
-        </Button>
+        <Content contentContainerStyle={{  margin: 25 }}>
+          <Card>
+            <CardItem>
+              <Text>Team Name: Team GTHC</Text>
+            </CardItem>
+            <CardItem>
+              <Text>Tent Type: Black</Text>
+            </CardItem>
+            <CardItem>
+              <Text>Passcode: ADC9L</Text>
+            </CardItem>
+          </Card>
+          <Button block style= {{marginTop: 40, margin: 50}}>
+            <Link to={'/editteamsettings'}>
+              <Text> Edit Team Settings </Text>
+            </Link>
+          </Button>
         </Content>
-        </Container>
+      </Container>
     );
   }
 }
@@ -70,4 +77,4 @@ const mapDispatchToProps = dispatch =>
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
-)(UserSettings);
+)(TeamSettings);
